requirements.txt
motor == 3.5.1
passlib == 1.7.4
python-jose == 3.3.0
pydantic == 2.8.2
uvicorn == 0.30.3